backend/requirements.txt
# FastAPI + Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Supabase
supabase==2.11.0

# Stripe
stripe==10.12.0

# Google Gemini AI
google-genai==1.5.0

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WhatsApp (Evolution API client)
requests==2.32.3
aiohttp==3.10.10
